% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{btc_labels}
\alias{btc_labels}
\title{Axis labels for breakthrough-curve figures (plotmath)}
\usage{
btc_labels()
}
\value{
list \code{conc}, \code{time} of plotmath expressions.
}
\description{
Units are taken from the data (the package's example data are in µg L⁻¹ and s).
}
