% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smim_forward.R
\name{lap_psi_ade}
\alias{lap_psi_ade}
\title{Laplace-domain exponential (ADE) waiting-time transform, \code{LapPsiFuncTPL.m} case 'ADE'}
\usage{
lap_psi_ade(s)
}
\description{
Laplace-domain exponential (ADE) waiting-time transform, \code{LapPsiFuncTPL.m} case 'ADE'
}
\keyword{internal}
