% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smim_forward.R
\name{lap_sol_tpl}
\alias{lap_sol_tpl}
\title{Laplace-domain SMIM solution (port of \code{LapSolTPL.m}, inlet/outlet BC both 'none')}
\usage{
lap_sol_tpl(s, transParams, options)
}
\arguments{
\item{s}{complex vector; @param transParams \code{c(vNorm, DNorm)};}

\item{options}{list: \verb{outletBCtype, inletBCfunc, xSample, inletBCtype, sorption params, "TPL", TPL params}.}
}
\description{
Laplace-domain SMIM solution (port of \code{LapSolTPL.m}, inlet/outlet BC both 'none')
}
\keyword{internal}
