% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smim_forward.R
\name{mem_func_tpl}
\alias{mem_func_tpl}
\title{Laplace-domain memory function (port of \code{memFuncTPL.m})}
\usage{
mem_func_tpl(s, options)
}
\description{
\code{options} mirrors the MATLAB cell: \code{list(c(lambda, tau, W, k_surf, k_sub), "TPL", c(beta, logT1, logT2))}.
}
\keyword{internal}
